+++ src/modules/admin/widgets/panels/FileHelpPanel.php
@@ -6,14 +6,14 @@ namespace davidhirtz\yii2\media\modules\admin\widgets\panels;
 
 use davidhirtz\yii2\media\assets\AdminAsset;
 use davidhirtz\yii2\media\models\File;
-use davidhirtz\yii2\media\modules\admin\widgets\forms\FileUploadInputWidget;
+use davidhirtz\yii2\media\modules\admin\widgets\buttons\ImportFileButton;
+use davidhirtz\yii2\media\modules\admin\widgets\buttons\UploadFileButton;
 use davidhirtz\yii2\media\modules\admin\widgets\panels\traits\DuplicateButtonTrait;
 use davidhirtz\yii2\media\modules\admin\widgets\panels\traits\FileLinkButtonTrait;
-use davidhirtz\yii2\skeleton\helpers\Html;
 use davidhirtz\yii2\skeleton\modules\admin\widgets\panels\HelpPanel;
+use Stringable;
 use Yii;
 use yii\helpers\Url;
-use yii\web\JsExpression;
 
 class FileHelpPanel extends HelpPanel
 {
@@ -50,38 +50,19 @@ class FileHelpPanel extends HelpPanel
         return $buttons;
     }
 
-    protected function getUploadFileButton(): string
+    protected function getUploadFileButton(): ?Stringable
     {
-        $content = Yii::t('media', 'Replace file');
-        $content = Html::iconText('upload', $content . $this->getFileUploadWidget());
-
-        return Html::tag('div', $content, [
-            'class' => 'btn btn-primary btn-upload',
-        ]);
+        return new UploadFileButton(
+            Yii::t('media', 'Replace file'),
+            Url::current(),
+        );
     }
 
-    protected function getImportFileButton(): string
+    protected function getImportFileButton(): ?Stringable
     {
-        return Html::tag('div', Html::iconText('cloud-upload-alt', Yii::t('media', 'Replace file')), [
-            'class' => 'btn btn-primary btn-submit btn-import',
-            'data' => [
-                'title' => Yii::t('media', 'Import file from URL'),
-                'url' => Url::current(),
-                'placeholder' => Yii::t('media', 'Link'),
-                'confirm' => Yii::t('media', 'Import'),
-            ],
-        ]);
-    }
-
-    protected function getFileUploadWidget(): string
-    {
-        return FileUploadInputWidget::widget([
-            'options' => [
-                'multiple' => false,
-            ],
-//            'clientEvents' => [
-//                'fileuploaddone' => new JsExpression('function(){location.reload();}')
-//            ],
-        ]);
+        return new ImportFileButton(
+            Yii::t('media', 'Replace file'),
+            Url::current(),
+        );
     }
 }