+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 3.0.0 (in development)
 
-- Removed `UploadTrait` in favor of `CopyFileButton` and `UploadFileButton` classes
+- Removed `UploadTrait` in favor of `ImportFileButton` and `UploadFileButton` classes
 
 ## 2.3.2 (Oct 20, 2025)